requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
gotrue==2.11.1
groq==0.15.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
postgrest==0.19.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.1.0
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
storage3==0.11.1
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
